+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-4.1.3 (unreleased)
+4.1.3 (2019-09-23)
 ------------------
 
 - Fixed bug "AttributeError: 'NoneType' object has no attribute 'lower'" in BaseDGHV.printAdvicesInfos when advice comment is None